Moved all python database connection info to a separate file: db_config_empty.py is the template, but needs to be copied to db_config.py and filled in. Don't check db_config.py into git. Made changes to other python scripts to use db_config. The main db build script checks to make sure db_config.py exists.

+++ dbscripts/build_dbexport.sh
@@ -7,6 +7,12 @@ DBEXPANSION="Odyssey 1.1.0"
 
 VERQUERY="INSERT INTO version VALUES ($DBVERSION,'$DBEXPANSION');"
 
+if [ ! -r db_config.py ]; then
+    echo "Missing db_config.py"
+    echo "Copy db_config_clean.py to db_config.py and fill in the database connection information"
+    exit
+fi
+
 rm -f tempdb.db
 rm -f rows.sql
